Code QoL improvements
