+++ src/Elastic.Markdown/Exporters/Elasticsearch/ElasticsearchIngestChannel.cs
@@ -1,0 +1,159 @@
+// Licensed to Elasticsearch B.V under one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information
+
+using Elastic.Channels;
+using Elastic.Documentation.Configuration;
+using Elastic.Documentation.Diagnostics;
+using Elastic.Documentation.Search;
+using Elastic.Documentation.Serialization;
+using Elastic.Ingest.Elasticsearch.Catalog;
+using Elastic.Ingest.Elasticsearch.Indices;
+using Elastic.Ingest.Elasticsearch.Semantic;
+using Elastic.Transport;
+using Microsoft.Extensions.Logging;
+
+namespace Elastic.Markdown.Exporters.Elasticsearch;
+
+public class ElasticsearchLexicalIngestChannel(
+	ILoggerFactory logFactory,
+	IDiagnosticsCollector collector,
+	ElasticsearchEndpoint endpoint,
+	string indexNamespace,
+	DistributedTransport transport,
+	string[] indexTimeSynonyms
+)
+	: ElasticsearchIngestChannel<CatalogIndexChannelOptions<DocumentationDocument>, CatalogIndexChannel<DocumentationDocument>>
+	(logFactory, collector, endpoint, transport, o => new(o), t => new(t)
+	{
+		BulkOperationIdLookup = d => d.Url,
+		// hash, last_updated and batch_index_date are all set before the docs are written to the channel
+		ScriptedHashBulkUpsertLookup = (d, _) => new HashedBulkUpdate("hash", d.Hash, "ctx._source.batch_index_date = params.batch_index_date",
+			new Dictionary<string, string>
+			{
+				{ "batch_index_date", d.BatchIndexDate.ToString("o") }
+			}),
+		GetMapping = () => CreateMapping(null),
+		GetMappingSettings = () => CreateMappingSetting($"docs-{indexNamespace}", indexTimeSynonyms),
+		IndexFormat =
+			$"{endpoint.IndexNamePrefix.Replace("semantic", "lexical").ToLowerInvariant()}-{indexNamespace.ToLowerInvariant()}-{{0:yyyy.MM.dd.HHmmss}}",
+		ActiveSearchAlias = $"{endpoint.IndexNamePrefix.Replace("semantic", "lexical").ToLowerInvariant()}-{indexNamespace.ToLowerInvariant()}"
+	});
+
+public class ElasticsearchSemanticIngestChannel(
+	ILoggerFactory logFactory,
+	IDiagnosticsCollector collector,
+	ElasticsearchEndpoint endpoint,
+	string indexNamespace,
+	DistributedTransport transport,
+	string[] indexTimeSynonyms
+)
+	: ElasticsearchIngestChannel<SemanticIndexChannelOptions<DocumentationDocument>, SemanticIndexChannel<DocumentationDocument>>
+	(logFactory, collector, endpoint, transport, o => new(o), t => new(t)
+	{
+		BulkOperationIdLookup = d => d.Url,
+		GetMapping = (inferenceId, _) => CreateMapping(inferenceId),
+		GetMappingSettings = (_, _) => CreateMappingSetting($"docs-{indexNamespace}", indexTimeSynonyms),
+		IndexFormat = $"{endpoint.IndexNamePrefix.ToLowerInvariant()}-{indexNamespace.ToLowerInvariant()}-{{0:yyyy.MM.dd.HHmmss}}",
+		ActiveSearchAlias = $"{endpoint.IndexNamePrefix}-{indexNamespace.ToLowerInvariant()}",
+		IndexNumThreads = endpoint.IndexNumThreads,
+		SearchNumThreads = endpoint.SearchNumThreads,
+		InferenceCreateTimeout = TimeSpan.FromMinutes(endpoint.BootstrapTimeout ?? 4),
+		UsePreexistingInferenceIds = !endpoint.NoElasticInferenceService,
+		InferenceId = endpoint.NoElasticInferenceService ? null : ".elser-2-elastic",
+		SearchInferenceId = endpoint.NoElasticInferenceService ? null : ".elser-2-elastic"
+	});
+
+public abstract partial class ElasticsearchIngestChannel<TChannelOptions, TChannel> : IDisposable
+	where TChannelOptions : CatalogIndexChannelOptionsBase<DocumentationDocument>
+	where TChannel : CatalogIndexChannel<DocumentationDocument, TChannelOptions>
+{
+	private readonly IDiagnosticsCollector _collector;
+	public TChannel Channel { get; }
+	private readonly ILogger _logger;
+
+	protected ElasticsearchIngestChannel(
+		ILoggerFactory logFactory,
+		IDiagnosticsCollector collector,
+		ElasticsearchEndpoint endpoint,
+		DistributedTransport transport,
+		Func<TChannelOptions, TChannel> createChannel,
+		Func<DistributedTransport, TChannelOptions> createOptions
+	)
+	{
+		_collector = collector;
+		_logger = logFactory.CreateLogger<ElasticsearchIngestChannel<TChannelOptions, TChannel>>();
+		//The max num threads per allocated node, from testing its best to limit our max concurrency
+		//producing to this number as well
+		var options = createOptions(transport);
+		var i = 0;
+		options.BufferOptions = new BufferOptions
+		{
+			OutboundBufferMaxSize = endpoint.BufferSize,
+			ExportMaxConcurrency = endpoint.IndexNumThreads,
+			ExportMaxRetries = endpoint.MaxRetries
+		};
+		options.SerializerContext = SourceGenerationContext.Default;
+		options.ExportBufferCallback = () =>
+		{
+			var count = Interlocked.Increment(ref i);
+			_logger.LogInformation("Exported {Count} documents to Elasticsearch index {IndexName}",
+				count * endpoint.BufferSize, Channel?.IndexName ?? string.Format(options.IndexFormat, "latest"));
+		};
+		options.ExportExceptionCallback = e =>
+		{
+			_logger.LogError(e, "Failed to export document");
+			_collector.EmitGlobalError("Elasticsearch export: failed to export document", e);
+		};
+		options.ServerRejectionCallback = items =>
+		{
+			foreach (var (doc, responseItem) in items)
+			{
+				_collector.EmitGlobalError(
+					$"Server rejection: {responseItem.Status} {responseItem.Error?.Type} {responseItem.Error?.Reason} for document {doc.Url}");
+			}
+		};
+		Channel = createChannel(options);
+		_logger.LogInformation("Created {Channel} Elasticsearch target for indexing", typeof(TChannel).Name);
+	}
+
+	public async ValueTask<bool> StopAsync(Cancel ctx = default)
+	{
+		_logger.LogInformation("Waiting to drain all inflight exports to Elasticsearch");
+		var drained = await Channel.WaitForDrainAsync(null, ctx);
+		if (!drained)
+			_collector.EmitGlobalError("Elasticsearch export: failed to complete indexing in a timely fashion while shutting down");
+
+		_logger.LogInformation("Refreshing target index {Index}", Channel.IndexName);
+		var refreshed = await Channel.RefreshAsync(ctx);
+		if (!refreshed)
+			_collector.EmitGlobalError($"Refreshing target index {Channel.IndexName} did not complete successfully");
+
+		_logger.LogInformation("Applying aliases to {Index}", Channel.IndexName);
+		var swapped = await Channel.ApplyAliasesAsync(ctx);
+		if (!swapped)
+			_collector.EmitGlobalError($"${nameof(ElasticsearchMarkdownExporter)} failed to apply aliases to index {Channel.IndexName}");
+
+		return drained && refreshed && swapped;
+	}
+
+	public async ValueTask<bool> RefreshAsync(Cancel ctx = default) => await Channel.RefreshAsync(ctx);
+
+	public async ValueTask<bool> TryWrite(DocumentationDocument document, Cancel ctx = default)
+	{
+		if (Channel.TryWrite(document))
+			return true;
+
+		if (await Channel.WaitToWriteAsync(ctx))
+			return Channel.TryWrite(document);
+		return false;
+	}
+
+	public void Dispose()
+	{
+		Channel.Complete();
+		Channel.Dispose();
+
+		GC.SuppressFinalize(this);
+	}
+}
