+++ tests-integration/Elastic.Assembler.IntegrationTests/Search/SearchBootstrapFixture.cs
@@ -176,7 +176,7 @@ public class SearchBootstrapFixture(DocumentationFixture fixture) : IAsyncLifeti
 			var collector = new ConsoleDiagnosticsCollector(loggerFactory);
 
 			// Create semantic exporter to check channel hash (index namespace is 'dev' for tests)
-			using var semanticExporter = new ElasticsearchSemanticExporter(
+			using var semanticExporter = new ElasticsearchSemanticIngestChannel(
 				loggerFactory,
 				collector,
 				endpoint,
